+++ CHANGELOG.md
@@ -11,3 +11,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Age range: removed description, added integer range
 - Dates: renamed to "reference_period", added "start" and
   "end" to all relevant datasets
+- Change resource metadata to provide format instead of MIME type